requirements.txt
asttokens==2.4.1
colorama==0.4.6
dnspython==2.4.2
executing==2.0.1
icecream==2.1.3
motor==3.3.1
Pygments==2.16.1
pymongo==4.5.0
python-dateutil==2.8.2
six==1.16.0

